Refactor flow C onboarding presentation into AltTextArticleEditorOnboardingPresenter
